+++ daemon/frontend/Frontend.h
@@ -67,7 +67,7 @@ protected:
 	void ServerSetDownloadRate(int rate);
 	bool RequestSetDownloadRate(int rate);
 	bool ServerEditQueue(DownloadQueue::EEditAction action, int offset, int entry);
-	bool RequestEditQueue(DownloadQueue::EEditAction action, int offset, int64 id);
+	bool RequestEditQueue(DownloadQueue::EEditAction action, int offset, uint32 id);
 	void Wait(int milliseconds);
 
 private:
